Add filter question dto
